pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "gallery"
version = "0.1"
dependencies = [
	"ffmpeg-python",
	"pillow",
	"pillow-heif",
	"exifread",
	"Flask"
]

[tool.setuptools.packages.find]
where = ["src"]
